flsk_stock_predict/requirements.txt
#astroid==2.9.3
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
black==21.12b0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.1
colorama==0.4.4
#config==0.5.1
coverage==6.3.1
cycler==0.10.0
Cython==0.29.24
decorator==5.1.1
executing==0.8.2
Flask==2.0.2
Flask-WTF==0.15.1
google-api-core==2.5.0
google-auth==2.6.0
google-cloud-bigquery==2.32.0
#google-cloud-core==2.2.2
google-cloud-storage==2.3.0
google-crc32c==1.3.0
#google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
grpcio==1.43.0
grpcio-status==1.43.0
##gunicorn==20.1.0
idna==3.3
iniconfig==1.1.1
ipython==8.0.0
isort==5.10.1
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
joblib==1.0.1
kiwisolver==1.3.2
lazy-object-proxy==1.7.1
MarkupSafe==2.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
numpy==1.21.2
packaging==21.3
pandas==1.3.3
parso==0.8.3
pathspec==0.9.0
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
platformdirs==2.4.1
pluggy==1.0.0
#pmdarima==1.8.3
prompt-toolkit==3.0.26
proto-plus==1.19.9
protobuf==3.19.4
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.11.2
pylint==2.12.2
pyparsing==2.4.7
#pystan==2.18.0.0
pytest==6.2.5
pytest-cov==3.0.0
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
rsa==4.8
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sklearn==0.0
stack-data==0.1.4
statsmodels==0.13.1
#tensorboard==2.1.0
#tensorflow==2.1.0
#tensorflow-estimator==2.1.0
threadpoolctl==3.0.0
toml==0.10.2
tomli==1.2.3
traitlets==5.1.1
typing_extensions==4.0.1
urllib3==1.26.7
wcwidth==0.2.5
Werkzeug==2.0.1
wrapt==1.13.3
WTForms==2.3.3
yfinance==0.1.70
